% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promis-package.R
\docType{package}
\name{promis-package}
\alias{promis}
\alias{promis-package}
\title{promis: Implementation of an Enhanced Vector Class for the PROMIS Health Measure}
\description{
TO DO.
}
\author{
\strong{Maintainer}: Nick Williams \email{niw4001@med.cornell.edu} (\href{https://orcid.org/0000-0002-1378-4831}{ORCID})

}
\keyword{internal}
