% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promis-package.R, R/promis.R
\name{promis-vctrs}
\alias{promis-vctrs}
\alias{vec_ptype2.promis_promis}
\alias{vec_cast.promis_promis}
\title{Internal vctrs methods}
\usage{
\method{vec_ptype2}{promis_promis}(x, y, ...)

\method{vec_cast}{promis_promis}(x, to, ...)
}
\description{
Internal vctrs methods
}
\keyword{internal}
