% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/promis.R
\name{promis}
\alias{promis}
\alias{is_promis}
\title{\code{promis} vector}
\usage{
promis(x = double(), domain = character())

is_promis(x)
}
\arguments{
\item{x}{\itemize{
\item For \code{promis()}: A numeric vector.
\item For \code{is_promis()}: A vector to test.
}}

\item{domain}{A character string specifying the PROMIS domain. Currently supported domains are:
\itemize{
\item Physical function
\item Pain interference
\item Pain intensity
\item Global physical health
\item Global mental health
\item Depression
}}
}
\value{
An S3 vector of class \code{promis_promis}.
}
\description{
\code{promis} vector
}
\examples{
promis(c(45, 60, 65), domain = "Physical function")
}
